requirements.txt
beautifulsoup4==4.8.0
Django==2.2.4
django-bootstrap4==1.0.1
django-crispy-forms==1.7.2
django-registration-redux==2.6
mysqlclient==1.4.4
paho-mqtt==1.4.0
Pillow==6.2.0
pytz==2019.2
soupsieve==1.9.3
sqlparse==0.3.0